Undo last commit and move Comment, Community, Person and Post api classes into api/base (to avoid circular references)
