+++ pylemmy/api/__init__.py
@@ -1,1 +1,1 @@
-from pylemmy.api import auth, comment, community2, listing, person, post, site  # noqa
+from pylemmy.api import auth, comment, community, listing, person, post, site  # noqa
