+++ pylemmy/api/person.py
@@ -1,33 +1,11 @@
 from typing import List, Optional
 
+from pylemmy.api.base import Person
 from pylemmy.api.comment import CommentView
-from pylemmy.api.community2 import CommunityModeratorView
+from pylemmy.api.community import CommunityModeratorView
 from pylemmy.api.listing import SortType
 from pylemmy.api.post import PostView
 from pylemmy.api.utils import BaseApiModel
-
-
-class Person(BaseApiModel):
-    actor_id: str
-    # admin: bool
-    # not to specification. Change to accept non-conforming API responses
-    admin: Optional[bool]
-    avatar: Optional[str]
-    ban_expires: Optional[str]
-    banned: bool
-    banner: Optional[str]
-    bio: Optional[str]
-    bot_account: bool
-    deleted: bool
-    display_name: Optional[str]
-    id: int
-    inbox_url: str
-    instance_id: int
-    local: bool
-    matrix_user_id: Optional[str]
-    name: str
-    published: str
-    updated: Optional[str]
 
 
 class PersonAggregates(BaseApiModel):
